After config options have been parsed a capturer plugin is created only if logging is to be captured.
Only a single handler and formatter are created and attached to the
root logger, instead of one for every test.

Each test has its own stream which the single handler is directed to
use for the duration of each test.

At the end of each test the stream is released so that it may be
garbage collected if appropriate.

Any logging that occurs outside of a test is output on the final
summary.