Refactoring and generalising the setting of log levels.
Renamed the plugin class.

Since there is now an implementation of a logging handler it handles both the log record instances
and the log text stream.

Factored out the funcarg into its own class rather than returning the
handler.

Factored out the log level context manager from the handler to it own class.

Store the handler on the item only since we can still get to it for the funcarg.

A reference to the handler was still on the function hence it would not be eligible for gc.  Now all
refs are removed after the test.

Removed remembering and restoring the root logger level after each test as it seemed like a noop.

Extended the use of separate checks for log, stdout and stderr to all tests.  In addition use the
py.test.raises checks for all tests.

Extended tests for setting log levels on any logger and also getting the log text from the funcarg.

Updated docs and made the README the authoritative one with the module docstring just a duplicate.